+++ tft_tban/CfgGroups.hpp
@@ -1,0 +1,207 @@
+class East {
+    class tft_tban {
+        name = "Taliban";
+        class Infantry {
+            name = "Infantry";
+
+            class Sentry {
+                side = 0;
+                faction = "Tban";
+                icon = "\Taliban_Fighters\data\cfgfactionclasses_Tban.paa";
+                name = "Sentry";
+                class Unit0 {
+                    side = 0;
+                    vehicle = "TBan_Rifleman2"; //R RPK
+                    rank = "SERGEANT";
+                    position[] = {0, 0, 0};
+                };
+                class Unit1 {
+                    side = 0;
+                    vehicle = "TBan_Rifleman"; //R
+                    rank = "CORPORAL";
+                    position[] = {5, -5, 0};
+                };
+            };
+            class Snipers {
+                side = 0;
+                faction = "Tban";
+                icon = "\Taliban_Fighters\data\cfgfactionclasses_Tban.paa";
+                name = "Sniper Team";
+                class Unit0 {
+                    side = 0;
+                    vehicle = "TBan_DMR"; //M
+                    rank = "SERGEANT";
+                    position[] = {0, 0, 0};
+                };
+                class Unit1 {
+                    side = 0;
+                    vehicle = "TBan_DMR"; //M
+                    rank = "CORPORAL";
+                    position[] = {5, -5, 0};
+                };
+            };
+            class Patrol {
+                side = 0;
+                faction = "Tban";
+                icon = "\Taliban_Fighters\data\cfgfactionclasses_Tban.paa";
+                name = "Patrol";
+                class Unit0 {
+                    side = 0;
+                    vehicle = "TBan_Rifleman2"; //R RPK
+                    rank = "SERGEANT";
+                    position[] = {0, 0, 0};
+                };
+                class Unit1 {
+                    side = 0;
+                    vehicle = "TBan_Rifleman"; //R
+                    rank = "CORPORAL";
+                    position[] = {5, -5, 0};
+                };
+                class Unit2 {
+                    side = 0;
+                    vehicle = "TBan_CLS"; //CLS
+                    rank = "PRIVATE";
+                    position[] = {-5, -5, 0};
+                };
+            };
+            class ATteam {
+                side = 0;
+                faction = "Tban";
+                icon = "\Taliban_Fighters\data\cfgfactionclasses_Tban.paa";
+                name = "AT Team";
+                class Unit0 {
+                    side = 0;
+                    vehicle = "TBan_Rifleman"; //R
+                    rank = "SERGEANT";
+                    position[] = {0, 0, 0};
+                };
+                class Unit1 {
+                    side = 0;
+                    vehicle = "TBan_AT"; //AT
+                    rank = "CORPORAL";
+                    position[] = {5, -5, 0};
+                };
+                class Unit2 {
+                    side = 0;
+                    vehicle = "TBan_AT"; //AT
+                    rank = "PRIVATE";
+                    position[] = {-5, -5, 0};
+                };
+            };
+            class Fireteam1 {
+                side = 0;
+                faction = "Tban";
+                icon = "\Taliban_Fighters\data\cfgfactionclasses_Tban.paa";
+                name = "Grunt Team";
+                class Unit0 {
+                    side = 0;
+                    vehicle = "TBan_TL"; //TL
+                    rank = "SERGEANT";
+                    position[] = {0, 0, 0};
+                };
+                class Unit1 {
+                    side = 0;
+                    vehicle = "TBan_Rifleman"; //R
+                    rank = "CORPORAL";
+                    position[] = {5, -5, 0};
+                };
+                class Unit2 {
+                    side = 0;
+                    vehicle = "TBan_Rifleman"; //R
+                    rank = "PRIVATE";
+                    position[] = {-5, -5, 0};
+                };
+                class Unit3 {
+                    side = 0;
+                    vehicle = "TBan_Rifleman2"; //R RPK
+                    rank = "PRIVATE";
+                    position[] = {10, -10, 0};
+                };
+                class Unit4 {
+                    side = 0;
+                    vehicle = "TBan_Rifleman2"; //R RPK
+                    rank = "PRIVATE";
+                    position[] = {-10, -10, 0};
+                };
+            };
+            class Fireteam2 {
+                side = 0;
+                faction = "Tban";
+                icon = "\Taliban_Fighters\data\cfgfactionclasses_Tban.paa";
+                name = "Fire Team";
+                class Unit0 {
+                    side = 0;
+                    vehicle = "TBan_TL"; //TL
+                    rank = "SERGEANT";
+                    position[] = {0, 0, 0};
+                };
+                class Unit1 {
+                    side = 0;
+                    vehicle = "TBan_AR"; //AR
+                    rank = "CORPORAL";
+                    position[] = {5, -5, 0};
+                };
+                class Unit2 {
+                    side = 0;
+                    vehicle = "TBan_CLS"; //CLS
+                    rank = "PRIVATE";
+                    position[] = {-5, -5, 0};
+                };
+                class Unit3 {
+                    side = 0;
+                    vehicle = "TBan_AT"; //AT
+                    rank = "PRIVATE";
+                    position[] = {10, -10, 0};
+                };
+                class Unit4 {
+                    side = 0;
+                    vehicle = "TBan_Rifleman"; //R
+                    rank = "PRIVATE";
+                    position[] = {-10, -10, 0};
+                };
+            };
+            class Fireteam3 {
+                side = 0;
+                faction = "Tban";
+                icon = "\Taliban_Fighters\data\cfgfactionclasses_Tban.paa";
+                name = "Heavy Team";
+                class Unit0 {
+                    side = 0;
+                    vehicle = "TBan_TL"; //TL
+                    rank = "SERGEANT";
+                    position[] = {0, 0, 0};
+                };
+                class Unit1 {
+                    side = 0;
+                    vehicle = "TBan_AR"; //AR
+                    rank = "CORPORAL";
+                    position[] = {5, -5, 0};
+                };
+                class Unit2 {
+                    side = 0;
+                    vehicle = "TBan_AR"; //AR
+                    rank = "PRIVATE";
+                    position[] = {-5, -5, 0};
+                };
+                class Unit3 {
+                    side = 0;
+                    vehicle = "TBan_AT"; //AT
+                    rank = "PRIVATE";
+                    position[] = {10, -10, 0};
+                };
+                class Unit4 {
+                    side = 0;
+                    vehicle = "TBan_DMR"; //M
+                    rank = "PRIVATE";
+                    position[] = {-10, -10, 0};
+                };
+                class Unit5 {
+                    side = 0;
+                    vehicle = "TBan_CLS"; //CLS
+                    rank = "PRIVATE";
+                    position[] = {15, -15, 0};
+                };
+            };
+        };
+    };
+};
